+++ public/knockout/kbSettings.js
@@ -9,30 +9,34 @@ function drawRow(device, scope, subScope) {
 	/* device */
 	row.append($("<td></td>"));
 	row.append($("<td>" + device.ip + "</td>"));
+	row.append($("<td>" + device.port + "</td>"));
 	row.append($("<td>" + device.name + "</td>"));
 	row.append($("<td>" + device.description + "</td>"));
 }
 
-function drawLocation(location, scope) {
+function drawLocation(obj, scope) {
 	var row = $("<tr />")
     $("#devices_table").append(row); //this will append tr element to table... keep its reference for a while since we will add cels into it
 
 	row.append($("<th scope=\"row\">"+ scope +"</th>"));
-	row.append($("<td>" + location.location + "</td>"));
+	row.append($("<td>" + obj.location + "</td>"));
 
 	/* device */
-	row.append($("<td>" + location.device[0].ip + "</td>"));
-	row.append($("<td>" + location.device[0].name + "</td>"));
-	row.append($("<td>" + location.device[0].description + "</td>"));
+	row.append($("<td>" + obj.devices[0].ip + "</td>"));
+	row.append($("<td>" + obj.devices[0].port + "</td>"));	
+	row.append($("<td>" + obj.devices[0].name + "</td>"));
+	row.append($("<td>" + obj.devices[0].description + "</td>"));
 }
 
 
-function drawTable(server){
+function drawTable(area){
 	/* location */
-	for (var i=0; i < server.length; i++) {
-		drawLocation(server[i],i);
-		for (var j=1; j< server[i].device.length; j++) {  
-			drawRow(server[i].device[j],i,j);
+	for (var i=0; i < area.length; i++) {
+		var a = area[i];
+		drawLocation(a,i);
+		for (var j=1; j< a.devices.length; j++) {
+			var dev = a.devices[j];
+			drawRow(dev,i,j);
 		}
 	}
 }
@@ -40,11 +44,10 @@ function drawTable(server){
 var ViewModel = function(ip) {
     
 	/* read json */
-	$.getJSON("/settings/config", function(data) { 
-		var config = data;
-    
-		if(config.light.server){
-			drawTable(config.light.server);    		
+	$.getJSON("/settings/config", function(json) {
+		var config = json;
+		if(config.light.area){
+			drawTable(config.light.area);    		
 		}else{
 			alert('invalid config - config.json');
 		}