+++ routes/menue.js
@@ -7,12 +7,24 @@ menue.get("/",function(req,res){
   res.sendFile(path.resolve(__dirname + "/../views/index.html"));
 });
 
-menue.get("/light",function(req,res){
-  res.sendFile(path.resolve(__dirname + "/../views/light.html"));
+menue.get("/remote",function(req,res){
+  res.sendFile(path.resolve(__dirname + "/../views/remote.html"));
 });
 
 menue.get("/settings",function(req,res){
   res.sendFile(path.resolve(__dirname + "/../views/settings.html"));
+});
+
+//menue.get("/settings/config",function(req,res){
+//	  res.sendFile(path.resolve(__dirname + "/../config.json"));
+//	});
+
+menue.get("/presets",function(req,res){
+	   res.sendFile(path.resolve(__dirname + "/../views/presets.html"));
+	});
+
+menue.get("/effects",function(req,res){
+	   res.sendFile(path.resolve(__dirname + "/../views/effects.html"));
 });
 
 menue.get("/test",function(req,res){