+++ routes/settings.js
@@ -4,12 +4,12 @@ var fs = require('fs');
 
 var path = require('path');
 
-settings.get("/settings/config",function(req,res){
+settings.get("/settings",function(req,res){
 	fs.readFile(__dirname + '/../config.json', function(err, data) {
   	if (err){
   		throw err;
   	}else{
-  		var json = JSON.parse(data); 				
+  		var json = JSON.parse(data);
 	    res.send(json);  	  
  	}
   });