+++ RELEASE_NOTES.md
@@ -1,6 +1,10 @@
 version 0.0.3
 ============
-  - add aurora API
+  - add effect - fork script	
+  - add presets to set color/element/pixel
+  - add range to remote to select 
+  	specific led's
+  - extract aurora API in own package
   
 version 0.0.2
 ============
@@ -14,5 +18,4 @@ version 0.0.2
 
 version 0.0.1
 ============
-
   - send tcp packages to one device